Test: 백오피스 예약건수+리스트 불러오기 통합 + 단위 테스트, PB 회원가입 테스트시 s3에 사진 업로드 되고 삭제하는 기능 추가

+++ src/main/java/kr/co/moneybridge/dto/backOffice/BackOfficeResponse.java
@@ -84,15 +84,15 @@ public class BackOfficeResponse {
     @Setter
     public static class ReservationTotalCountDTO {
         @ApiModelProperty(example = "2", value = "총 상담 신청 건수")
-        private Integer apply;
+        private Long apply;
         @ApiModelProperty(example = "3", value = "총 상담 확정 건수")
-        private Integer confirm;
+        private Long confirm;
         @ApiModelProperty(example = "2", value = "총 상담 완료 건수")
-        private Integer complete;
+        private Long complete;
         @ApiModelProperty(example = "2", value = "총 후기 작성 건수")
         private Long review;
 
-        public ReservationTotalCountDTO(Integer apply, Integer confirm, Integer complete, Long review) {
+        public ReservationTotalCountDTO(Long apply, Long confirm, Long complete, Long review) {
             this.apply = apply;
             this.confirm = confirm;
             this.complete = complete;
